Remove unecessary memory usage
member.py assigned a variable to a rvalue that is only ever checked
once, directly below it. It did nothing to cut down on the line length,
in fact it created a new, longer line. Readability is now supplied
through comments, and memory is not wasted every time that
signed_packets() is ran